add generateBridgeUiConfig to sdk
