+++ src/types/bridgeUiConfig.ts
@@ -1,0 +1,67 @@
+import { Chain, PublicClient, Transport } from "viem";
+
+/**
+ * Bridge UI configuration information for adding a new test arbitrum chain
+ */
+export type BridgeUiConfig = {
+    /**
+     * Basic chain information
+     */
+    chainInfo: {
+      chainName: string;
+      chainId: number;
+      parentChainId: number;
+      rpcUrl: string;
+      explorerUrl: string;
+      nativeToken?: string;
+    };
+    /**
+     * Core protocol contract addresses
+     */
+    coreContracts: {
+      rollup: string;
+      inbox: string;
+      outbox: string;
+      sequencerInbox: string;
+      bridge: string;
+      nativeToken?: string;
+    };
+    /**
+     * Token bridge contract addresses for both parent chain (L2) and Orbit chain (L3)
+     */
+    tokenBridgeContracts: {
+      /** Parent chain (L2) contract addresses */
+      l2Contracts: {
+        customGateway: string;
+        multicall: string;
+        proxyAdmin: string;
+        router: string;
+        standardGateway: string;
+        weth: string;
+        wethGateway: string;
+      };
+      /** Orbit chain (L3) contract addresses */
+      l3Contracts: {
+        customGateway: string;
+        multicall: string;
+        proxyAdmin: string;
+        router: string;
+        standardGateway: string;
+        weth: string;
+        wethGateway: string;
+      };
+    };
+}
+
+
+/**
+ * Parameters for retrieving bridge UI configuration
+ */
+export type GetBridgeUiConfigParams<TChain extends Chain | undefined> = {
+    parentChain: TChain;
+    deploymentTxHash: `0x${string}`;
+    chainName?: string;
+    rpcUrl?: string;
+    explorerUrl?: string;
+    parentChainPublicClient?: PublicClient<Transport, TChain>;
+}
